test scraps 2
Testing so far remains pretty manual, partly because
setting up automated testing that works across all
of the supported Pythons is its own portability
challenge. But this is another move in the direction
of proper automated tests once day.

Another move is getting to a point where I can run
all of the browser Pythons in the same environment.

* Skulpt:
  1. `npm run build` possibly with the environment variable
     `NODE_OPTIONS=--openssl-legacy-provider`, then
  2. `npm run repl (py2|py3)`, but still need to figure out
     how to reliably feed Python into this, or alternatively
     how to run things without going through that REPL.
* Transcrypt:
  1. build with `transcrypt` (`-n` option avoids the Java
     minifier dependency, and `-b` to rebuild everything
     each time might be useful.), and
  2. use modules in __target__ with Node (or maybe Deno?),
     possibly after adjusting imports + renaming .js->.mjs.
* Brython: only briefly looked into this one, not sure if
  there is any official or polished way to run it locally.

For contrast, Pyodide just uses a browser through Selenium
for tests, and maybe that's the way to go.